# docs/openapi-v1-incremental.yaml
openapi: 3.1.0
info:
  title: 淘到宝引擎 V1 增量契约
  version: 1.1.0
  description: >-
    本文件只描述相对冻结 V0.5 openapi.yaml 新增的接口；冻结路径和字段语义保持不变。
servers:
  - url: /api/v1
paths:
...
  /feishu/resources:
    get:
      operationId: listFeishuResources
      parameters:
        - name: type
          in: query
          required: true
          schema: {enum: [document, minute]}
        - name: page_token
          in: query
          schema: {type: [string, 'null']}
      responses:
        '200': {description: 当前用户有权限访问的飞书文档或妙记列表}
  /feishu/bitable/sync-daily:
    post:
      operationId: syncDailyWorkspaceToFeishuBitable
      description: >-
        将当前工作区的客户画像、资料来源、已校验经验、原子能力、招标项目与
        Deep Research 状态单向写入当前登录用户的真实飞书多维表格。首次调用
        创建“淘到宝·今日战情”，后续调用复用同一 app_token/table_id；该接口
        不反向覆盖长期画像或可信资产，也不产生 Mock 业务记录。
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '200':
          description: 多维表格同步完成
          content:
            application/json:
              schema:
                type: object
                required: [request_id, data]
                properties:
                  request_id: {type: string}
                  data:
                    type: object
                    required: [status, created, record_count, url, synced_at]
                    properties:
                      status: {const: synced}
                      created: {type: boolean}
                      record_count: {type: integer, minimum: 0}
                      url: {type: string}
                      app_token: {type: string}
                      table_id: {type: string}
                      synced_at: {type: string, format: date-time}
        '503': {description: 飞书权限、授权状态或 OpenAPI 暂不可用}
...
  /expert-collaborations:
    post:
      operationId: createExpertCollaboration
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [research_task_id]
              properties:
                research_task_id: {type: string, format: uuid}
      responses:
        '201': {$ref: '#/components/responses/ExpertCollaborationResponse'}
    get:
      operationId: listExpertCollaborations
      responses:
        '200': {description: 当前工作区专家协作分页列表}
  /expert-collaborations/{collaboration_id}:
    parameters: [{$ref: '#/components/parameters/CollaborationId'}]
    get:
      operationId: getExpertCollaboration
      responses:
        '200': {$ref: '#/components/responses/ExpertCollaborationResponse'}
    patch:
      operationId: updateExpertCollaboration
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateExpertCollaborationRequest'}
      responses:
        '200': {$ref: '#/components/responses/ExpertCollaborationResponse'}
  /expert-collaborations/{collaboration_id}/confirm:
    parameters: [{$ref: '#/components/parameters/CollaborationId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
    post:
      operationId: confirmExpertCollaboration
      responses:
        '200': {$ref: '#/components/responses/ExpertCollaborationResponse'}
  /expert-collaborations/events/replies:
    post:
      operationId: receiveExpertReply
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpertReplyEvent'}
      responses:
        '202': {description: 去重保存回复并恢复原研究任务}
  /expert-collaborations/events/feishu:
    post:
      operationId: receiveNativeFeishuEvent
      description: >-
        飞书开放平台原生事件回调。支持 url_verification challenge、加密事件、
        X-Lark-Signature 校验及 im.message.receive_v1 群文本消息；此入口无需业务登录态。
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        '200': {description: challenge 响应、忽略非目标事件或接收成功}
        '401': {description: verification token、签名或加密数据校验失败}
  /expert-collaborations/replies/{reply_id}/adopt:
    parameters:
      - {name: reply_id, in: path, required: true, schema: {type: string, format: uuid}}
      - {$ref: '#/components/parameters/IdempotencyKey'}
    post:
      operationId: adoptExpertReply
      responses:
        '200': {description: 创建 pending_review 经验草稿，绝不自动校验}
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128}
...
    CollaborationId:
      name: collaboration_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  schemas:
...
    UpdateExpertCollaborationRequest:
      type: object
      required: [group_name, selected_expert_ids, questions]
      properties:
        group_name: {type: string, minLength: 1, maxLength: 200}
        selected_expert_ids: {type: array, minItems: 1, maxItems: 3, items: {type: string, format: uuid}}
        questions: {type: array, minItems: 1, maxItems: 8, items: {type: object}}
    ExpertReplyEvent:
      type: object
      required: [collaboration_id, question_id, author_id, author_name, answer_text, feishu_message_id, message_url]
      properties:
        collaboration_id: {type: string, format: uuid}
        question_id: {type: string}
        author_id: {type: string}
        author_name: {type: string}
        answer_text: {type: string}
        feishu_message_id: {type: string}
        message_url: {type: string}
        verification_token: {type: [string, 'null']}
  responses:
...
    ExpertCollaborationResponse:
      description: 专家协作响应信封